+++ src/common/collections/tenants/tabular.js
@@ -99,10 +99,10 @@ Tracker.autorun(() => {
             },
             // the publication takes care of providing the list of fields which have not the same value among all records
             createdRow( row, data, dataIndex, cells ){
-                //console.debug( data );
+                //console.debug( columns, data );
                 // set a different display when a value changes between validity records
                 data.DYN.analyze.diffs.forEach(( it ) => {
-                    const def = columns[it].def;
+                    const def = columns[it]?.def;
                     if( def && def.dt_tabular !== false && def.dt_visible !== false ){
                         $( cells[columns[it].index] ).addClass( 'dt-different' );   
                     }