+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - Open the new tenant dialog on first pane
     - Define new 'scopedManagerRole' configuration parameter (todo #4)
     - Fix the tabular's row management when the column is not displayed
+    - Define new 'tabularServerExtend' configuration parameter
 
 ### 1.2.0