+++ ChangeLog.md
@@ -20,6 +20,7 @@
     - Extends and maintains TenantsManager.list on server side
     - Fix default ordering by ascending email address
     - Improve (and fix) the display of validity periods count
+    - Define new 'serverAllExtend' configuration parameter
 
 ### 1.2.0